lpc: phytium: Add LPC snoop driver
The LPC snoop driver is used to monitor the startup progress of the CPU
side Server. Generally,port 0x80 is used.

The usage is as follows: The firmware writes the current system
startup progress to port 0x80 during the server CPU side startup,
and then the system has an APP program to obtain the data on port
0x80 from the snoop and output it to the terminal.

Signed-off-by: Li Yuting <liyuting2071@phytium.com.cn>
Signed-off-by: Lan Hengyu <lanhengyu1395@phytium.com.cn>
Signed-off-by: Wang Yinfeng <wangyinfeng@phytium.com.cn>